+++ test/blackbox/types/statistics/typesPubSubTypes.cxx
@@ -40,7 +40,7 @@ namespace eprosima {
                     setName("eprosima::fastdds::statistics::detail::EntityId_s");
                     uint32_t type_size =
                 #if FASTCDR_VERSION_MAJOR == 1
-                        EntityId_s::getMaxCdrSerializedSize();
+                        static_cast<uint32_t>(EntityId_s::getMaxCdrSerializedSize());
                 #else
                         eprosima_fastdds_statistics_detail_EntityId_s_max_cdr_typesize;
                 #endif
@@ -230,7 +230,7 @@ namespace eprosima {
                     setName("eprosima::fastdds::statistics::detail::GuidPrefix_s");
                     uint32_t type_size =
                 #if FASTCDR_VERSION_MAJOR == 1
-                        GuidPrefix_s::getMaxCdrSerializedSize();
+                        static_cast<uint32_t>(GuidPrefix_s::getMaxCdrSerializedSize());
                 #else
                         eprosima_fastdds_statistics_detail_GuidPrefix_s_max_cdr_typesize;
                 #endif
@@ -420,7 +420,7 @@ namespace eprosima {
                     setName("eprosima::fastdds::statistics::detail::GUID_s");
                     uint32_t type_size =
                 #if FASTCDR_VERSION_MAJOR == 1
-                        GUID_s::getMaxCdrSerializedSize();
+                        static_cast<uint32_t>(GUID_s::getMaxCdrSerializedSize());
                 #else
                         eprosima_fastdds_statistics_detail_GUID_s_max_cdr_typesize;
                 #endif
@@ -610,7 +610,7 @@ namespace eprosima {
                     setName("eprosima::fastdds::statistics::detail::SequenceNumber_s");
                     uint32_t type_size =
                 #if FASTCDR_VERSION_MAJOR == 1
-                        SequenceNumber_s::getMaxCdrSerializedSize();
+                        static_cast<uint32_t>(SequenceNumber_s::getMaxCdrSerializedSize());
                 #else
                         eprosima_fastdds_statistics_detail_SequenceNumber_s_max_cdr_typesize;
                 #endif
@@ -800,7 +800,7 @@ namespace eprosima {
                     setName("eprosima::fastdds::statistics::detail::SampleIdentity_s");
                     uint32_t type_size =
                 #if FASTCDR_VERSION_MAJOR == 1
-                        SampleIdentity_s::getMaxCdrSerializedSize();
+                        static_cast<uint32_t>(SampleIdentity_s::getMaxCdrSerializedSize());
                 #else
                         eprosima_fastdds_statistics_detail_SampleIdentity_s_max_cdr_typesize;
                 #endif
@@ -990,7 +990,7 @@ namespace eprosima {
                     setName("eprosima::fastdds::statistics::detail::Locator_s");
                     uint32_t type_size =
                 #if FASTCDR_VERSION_MAJOR == 1
-                        Locator_s::getMaxCdrSerializedSize();
+                        static_cast<uint32_t>(Locator_s::getMaxCdrSerializedSize());
                 #else
                         eprosima_fastdds_statistics_detail_Locator_s_max_cdr_typesize;
                 #endif
@@ -1183,7 +1183,7 @@ namespace eprosima {
                 setName("eprosima::fastdds::statistics::DiscoveryTime");
                 uint32_t type_size =
             #if FASTCDR_VERSION_MAJOR == 1
-                    DiscoveryTime::getMaxCdrSerializedSize();
+                    static_cast<uint32_t>(DiscoveryTime::getMaxCdrSerializedSize());
             #else
                     eprosima_fastdds_statistics_DiscoveryTime_max_cdr_typesize;
             #endif
@@ -1373,7 +1373,7 @@ namespace eprosima {
                 setName("eprosima::fastdds::statistics::EntityCount");
                 uint32_t type_size =
             #if FASTCDR_VERSION_MAJOR == 1
-                    EntityCount::getMaxCdrSerializedSize();
+                    static_cast<uint32_t>(EntityCount::getMaxCdrSerializedSize());
             #else
                     eprosima_fastdds_statistics_EntityCount_max_cdr_typesize;
             #endif
@@ -1563,7 +1563,7 @@ namespace eprosima {
                 setName("eprosima::fastdds::statistics::SampleIdentityCount");
                 uint32_t type_size =
             #if FASTCDR_VERSION_MAJOR == 1
-                    SampleIdentityCount::getMaxCdrSerializedSize();
+                    static_cast<uint32_t>(SampleIdentityCount::getMaxCdrSerializedSize());
             #else
                     eprosima_fastdds_statistics_SampleIdentityCount_max_cdr_typesize;
             #endif
@@ -1753,7 +1753,7 @@ namespace eprosima {
                 setName("eprosima::fastdds::statistics::Entity2LocatorTraffic");
                 uint32_t type_size =
             #if FASTCDR_VERSION_MAJOR == 1
-                    Entity2LocatorTraffic::getMaxCdrSerializedSize();
+                    static_cast<uint32_t>(Entity2LocatorTraffic::getMaxCdrSerializedSize());
             #else
                     eprosima_fastdds_statistics_Entity2LocatorTraffic_max_cdr_typesize;
             #endif
@@ -1943,7 +1943,7 @@ namespace eprosima {
                 setName("eprosima::fastdds::statistics::WriterReaderData");
                 uint32_t type_size =
             #if FASTCDR_VERSION_MAJOR == 1
-                    WriterReaderData::getMaxCdrSerializedSize();
+                    static_cast<uint32_t>(WriterReaderData::getMaxCdrSerializedSize());
             #else
                     eprosima_fastdds_statistics_WriterReaderData_max_cdr_typesize;
             #endif
@@ -2133,7 +2133,7 @@ namespace eprosima {
                 setName("eprosima::fastdds::statistics::Locator2LocatorData");
                 uint32_t type_size =
             #if FASTCDR_VERSION_MAJOR == 1
-                    Locator2LocatorData::getMaxCdrSerializedSize();
+                    static_cast<uint32_t>(Locator2LocatorData::getMaxCdrSerializedSize());
             #else
                     eprosima_fastdds_statistics_Locator2LocatorData_max_cdr_typesize;
             #endif
@@ -2323,7 +2323,7 @@ namespace eprosima {
                 setName("eprosima::fastdds::statistics::EntityData");
                 uint32_t type_size =
             #if FASTCDR_VERSION_MAJOR == 1
-                    EntityData::getMaxCdrSerializedSize();
+                    static_cast<uint32_t>(EntityData::getMaxCdrSerializedSize());
             #else
                     eprosima_fastdds_statistics_EntityData_max_cdr_typesize;
             #endif
@@ -2513,7 +2513,7 @@ namespace eprosima {
                 setName("eprosima::fastdds::statistics::PhysicalData");
                 uint32_t type_size =
             #if FASTCDR_VERSION_MAJOR == 1
-                    PhysicalData::getMaxCdrSerializedSize();
+                    static_cast<uint32_t>(PhysicalData::getMaxCdrSerializedSize());
             #else
                     eprosima_fastdds_statistics_PhysicalData_max_cdr_typesize;
             #endif
